fix: pipewire destroy issue introduced by synchronization changes
This was an issue at one point, but then it was resolved. If
pipewire is screencasting a screen and there are no changes,
it the channels sending to the record thread don't trigger,
causing it to hang. We need to make sure there are no pipewire
channels hung up when the screencast is stopped.